+++ persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/News.java
@@ -35,7 +35,7 @@ public class News implements Serializable {
 
 	@XmlTransient
 	@Column(name="community", insertable=false, updatable=false)
-	private byte communityId;
+	private Integer communityId;
 	
 	@ManyToOne(fetch=FetchType.LAZY)
 	@JoinColumn(name="community")
@@ -77,7 +77,7 @@ public class News implements Serializable {
 		communityId=community.getId();
 	}
 	
-	public byte getCommunityId() {
+	public Integer getCommunityId() {
 		return communityId;
 	}
